requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0.1
Pillow==7.2.0
psycopg2-binary==2.9.5

# Django
django==3.1.1

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.3.0

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1
tornado>=5.0.0,<6.0.0
flower==0.9.2
